Tidy a spectra collection

tidy will return a dataframe containing the spectral information in its current state.

# S3 method for collection
tidy(x, ...)

Arguments

x
A collection object.

...
further arguments passed to or from other methods(not currenctly used).

Value

A tibble with a variable number columns depending on the binning of the data:

id

Unique sample identifier

label

Label for the sample

bins

Bin centers

Examples

library(tidySpectR)

tidy(fa_nmr)

#> # A tibble: 6 x 2,252
#>   id    label `0.001978768338… `0.005978768338… `0.009978768338…
#>   <fct> <fct>            <dbl>            <dbl>            <dbl>
#> 1 2019… conv…         2591796.         2853272          2852076.
#> 2 2019… conv…         2416512.         2613998.         2689199 
#> 3 2019… conv…         1965228.         2067044.         2080362 
#> 4 2019… orga…         2576618.         2653498.         2730184 
#> 5 2019… orga…         2335314          2401902          2421192 
#> 6 2019… orga…         2415644.         2510800          2504370 
#> # … with 2,247 more variables: `0.0139787683387494` <dbl>,
#> #   `0.0179787683387494` <dbl>, `0.0219787683387494` <dbl>,
#> #   `0.0259787683387494` <dbl>, `0.0299787683387494` <dbl>,
#> #   `0.0339787683387494` <dbl>, `0.0379787683387494` <dbl>,
#> #   `0.0419787683387494` <dbl>, `0.0459787683387494` <dbl>,
#> #   `0.0499787683387494` <dbl>, `0.0539787683387494` <dbl>,
#> #   `0.0579787683387494` <dbl>, `0.0619787683387494` <dbl>,
#> #   `0.0659787683387494` <dbl>, `0.0699787683387494` <dbl>,
#> #   `0.0739787683387494` <dbl>, `0.0779787683387494` <dbl>,
#> #   `0.0819787683387494` <dbl>, `0.0859787683387494` <dbl>,
#> #   `0.0899787683387494` <dbl>, `0.0939787683387494` <dbl>,
#> #   `0.0979787683387494` <dbl>, `0.101978768338749` <dbl>,
#> #   `0.105978768338749` <dbl>, `0.109978768338749` <dbl>,
#> #   `0.113978768338749` <dbl>, `0.117978768338749` <dbl>,
#> #   `0.121978768338749` <dbl>, `0.125978768338749` <dbl>,
#> #   `0.129978768338749` <dbl>, `0.133978768338749` <dbl>,
#> #   `0.137978768338749` <dbl>, `0.141978768338749` <dbl>,
#> #   `0.145978768338749` <dbl>, `0.149978768338749` <dbl>,
#> #   `0.153978768338749` <dbl>, `0.157978768338749` <dbl>,
#> #   `0.161978768338749` <dbl>, `0.165978768338749` <dbl>,
#> #   `0.169978768338749` <dbl>, `0.173978768338749` <dbl>,
#> #   `0.177978768338749` <dbl>, `0.181978768338749` <dbl>,
#> #   `0.185978768338749` <dbl>, `0.189978768338749` <dbl>,
#> #   `0.193978768338749` <dbl>, `0.197978768338749` <dbl>,
#> #   `0.201978768338749` <dbl>, `0.205978768338749` <dbl>,
#> #   `0.209978768338749` <dbl>, `0.213978768338749` <dbl>,
#> #   `0.217978768338749` <dbl>, `0.221978768338749` <dbl>,
#> #   `0.225978768338749` <dbl>, `0.229978768338749` <dbl>,
#> #   `0.233978768338749` <dbl>, `0.237978768338749` <dbl>,
#> #   `0.241978768338749` <dbl>, `0.245978768338749` <dbl>,
#> #   `0.249978768338749` <dbl>, `0.253978768338749` <dbl>,
#> #   `0.257978768338749` <dbl>, `0.261978768338749` <dbl>,
#> #   `0.265978768338749` <dbl>, `0.269978768338749` <dbl>,
#> #   `0.273978768338749` <dbl>, `0.277978768338749` <dbl>,
#> #   `0.281978768338749` <dbl>, `0.285978768338749` <dbl>,
#> #   `0.289978768338749` <dbl>, `0.293978768338749` <dbl>,
#> #   `0.297978768338749` <dbl>, `0.301978768338749` <dbl>,
#> #   `0.305978768338749` <dbl>, `0.309978768338749` <dbl>,
#> #   `0.313978768338749` <dbl>, `0.317978768338749` <dbl>,
#> #   `0.321978768338749` <dbl>, `0.325978768338749` <dbl>,
#> #   `0.329978768338749` <dbl>, `0.333978768338749` <dbl>,
#> #   `0.337978768338749` <dbl>, `0.341978768338749` <dbl>,
#> #   `0.345978768338749` <dbl>, `0.349978768338749` <dbl>,
#> #   `0.353978768338749` <dbl>, `0.357978768338749` <dbl>,
#> #   `0.361978768338749` <dbl>, `0.365978768338749` <dbl>,
#> #   `0.369978768338749` <dbl>, `0.373978768338749` <dbl>,
#> #   `0.377978768338749` <dbl>, `0.381978768338749` <dbl>,
#> #   `0.385978768338749` <dbl>, `0.389978768338749` <dbl>,
#> #   `0.393978768338749` <dbl>, `0.397978768338749` <dbl>,
#> #   `0.401978768338749` <dbl>, `0.405978768338749` <dbl>,
#> #   `0.409978768338749` <dbl>, …