Tidy a spectra collection

tidy will return a dataframe containing the spectral information in its current state.

# S3 method for collection
tidy(x, ...)

Arguments

x
A collection object.

...
further arguments passed to or from other methods(not currenctly used).

Value

A tibble with a variable number columns depending on the binning of the data:

id

Unique sample identifier

label

Label for the sample

bins

Bin centers

Examples

library(tidySpectR)

tidy(fa_nmr)

#> # A tibble: 10 x 47,502
#>    id    label `-4.9998` `-4.9994` `-4.999` `-4.9986` `-4.9982` `-4.9978`
#>    <fct> <fct>     <dbl>     <dbl>    <dbl>     <dbl>     <dbl>     <dbl>
#>  1 2019… cont…   -13743    -14571   -17018    -11326.   -18380    -24434.
#>  2 2019… cont…   -13460     -5015    -9891    -13904.   -18492.   -22658 
#>  3 2019… cont…   -19258    -18456.  -24750    -22734    -27160.   -26340.
#>  4 2019… cont…   -24185    -20839   -19651     -8882    -15314    -20772.
#>  5 2019… cont…   -18325    -21071   -20296.   -10256.    -9047     -8506 
#>  6 2019… trea…   -23787    -26620.  -31168.   -18542.   -17213    -24420 
#>  7 2019… trea…    -9260.   -19420   -27654    -21138.   -13868.   -16014 
#>  8 2019… trea…   -13450.   -14439   -24902.   -28344.   -29628    -22702 
#>  9 2019… trea…   -12128.   -18972.  -23252    -21679    -26698.   -23290 
#> 10 2019… trea…   -19702.   -21337   -20188     -6712.   -14839    -27977 
#> # … with 47,494 more variables: `-4.9974` <dbl>, `-4.997` <dbl>,
#> #   `-4.9966` <dbl>, `-4.9962` <dbl>, `-4.9958` <dbl>, `-4.9954` <dbl>,
#> #   `-4.995` <dbl>, `-4.9946` <dbl>, `-4.9942` <dbl>, `-4.9938` <dbl>,
#> #   `-4.9934` <dbl>, `-4.993` <dbl>, `-4.9926` <dbl>, `-4.9922` <dbl>,
#> #   `-4.9918` <dbl>, `-4.9914` <dbl>, `-4.991` <dbl>, `-4.9906` <dbl>,
#> #   `-4.9902` <dbl>, `-4.9898` <dbl>, `-4.9894` <dbl>, `-4.989` <dbl>,
#> #   `-4.9886` <dbl>, `-4.9882` <dbl>, `-4.9878` <dbl>, `-4.9874` <dbl>,
#> #   `-4.987` <dbl>, `-4.9866` <dbl>, `-4.9862` <dbl>, `-4.9858` <dbl>,
#> #   `-4.9854` <dbl>, `-4.985` <dbl>, `-4.9846` <dbl>, `-4.9842` <dbl>,
#> #   `-4.9838` <dbl>, `-4.9834` <dbl>, `-4.983` <dbl>, `-4.9826` <dbl>,
#> #   `-4.9822` <dbl>, `-4.9818` <dbl>, `-4.9814` <dbl>, `-4.981` <dbl>,
#> #   `-4.9806` <dbl>, `-4.9802` <dbl>, `-4.9798` <dbl>, `-4.9794` <dbl>,
#> #   `-4.979` <dbl>, `-4.9786` <dbl>, `-4.9782` <dbl>, `-4.9778` <dbl>,
#> #   `-4.9774` <dbl>, `-4.977` <dbl>, `-4.9766` <dbl>, `-4.9762` <dbl>,
#> #   `-4.9758` <dbl>, `-4.9754` <dbl>, `-4.975` <dbl>, `-4.9746` <dbl>,
#> #   `-4.9742` <dbl>, `-4.9738` <dbl>, `-4.9734` <dbl>, `-4.973` <dbl>,
#> #   `-4.9726` <dbl>, `-4.9722` <dbl>, `-4.9718` <dbl>, `-4.9714` <dbl>,
#> #   `-4.971` <dbl>, `-4.9706` <dbl>, `-4.9702` <dbl>, `-4.9698` <dbl>,
#> #   `-4.9694` <dbl>, `-4.969` <dbl>, `-4.9686` <dbl>, `-4.9682` <dbl>,
#> #   `-4.9678` <dbl>, `-4.9674` <dbl>, `-4.967` <dbl>, `-4.9666` <dbl>,
#> #   `-4.9662` <dbl>, `-4.9658` <dbl>, `-4.9654` <dbl>, `-4.965` <dbl>,
#> #   `-4.9646` <dbl>, `-4.9642` <dbl>, `-4.9638` <dbl>, `-4.9634` <dbl>,
#> #   `-4.963` <dbl>, `-4.9626` <dbl>, `-4.9622` <dbl>, `-4.9618` <dbl>,
#> #   `-4.9614` <dbl>, `-4.961` <dbl>, `-4.9606` <dbl>, `-4.9602` <dbl>,
#> #   `-4.9598` <dbl>, `-4.9594` <dbl>, `-4.959` <dbl>, `-4.9586` <dbl>,
#> #   `-4.9582` <dbl>, `-4.9578` <dbl>, …